06. Conexión a base de datos, crear, editar y eliminar contenido

+++ FundamentosCSHARP/Models/BebidaAlcoholica.cs
@@ -8,7 +8,7 @@ namespace FundamentosCSHARP.Models
 {
     interface IBebidaAlcoholica
     {
-        public int Alcohol { get; set; }
+        public decimal Alcohol { get; set; }
 
         public void MaxRecomendado();
     }
