+++ FundamentosCSHARP/Models/Cerveza.cs
@@ -8,7 +8,7 @@ namespace FundamentosCSHARP.Models
 {
     class Cerveza : Bebida, IBebidaAlcoholica
     {
-        public int Alcohol { get; set; }
+        public decimal Alcohol { get; set; }
         public string Marca { get; set; }
         public void MaxRecomendado() 
         {
